Safari-609.1.11.1/LayoutTests/media/controls/fullscreen-button-inline-layout-expected.txt

This tests for proper flexbox layout by checking the location of the fullscreen button. It should be within the bounds of the inline bar instead of being pushed past the right edge.

This test only runs in DRT!


EVENT: canplaythrough
PASS: We are using the apple idiom
PASS: Fullscreen button is placed within bounds of inline controls

Testing finished.
